backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
openai==1.12.0
qdrant-client==1.7.0
sentence-transformers==2.3.1
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6
numpy==1.26.3

